+++ src/includes/AlmaSettings.php
@@ -72,6 +72,7 @@ use Alma\Woocommerce\Helpers\PlanBuilderHelper;
  * @property string share_of_checkout_enabled_date String Date when the merchant did accept the share of checkout
  * @property string share_of_checkout_last_sharing_date String Date when we sent the data to Alma
  * @property bool display_in_page Bool if In Page is activated
+ * @property bool use_blocks_template Bool if we want to use a blocks template
  */
 class AlmaSettings {
 
@@ -165,12 +166,12 @@ class AlmaSettings {
 	}
 
 	/**
-	 * Is logging enabled.
+	 * Is blocks template enabled.
 	 *
 	 * @return bool
 	 */
-	public function is_logging_enabled() {
-		return 'yes' === $this->debug;
+	public function is_blocks_template_enabled() {
+		return 'yes' === $this->use_blocks_template;
 	}
 
 	/**
@@ -383,10 +384,15 @@ class AlmaSettings {
 	 * Gets title for a payment method.
 	 *
 	 * @param string $payment_method The payment method.
+	 * @param string $is_blocks Are we in blocks.
 	 *
 	 * @return string
 	 */
-	public function get_title( $payment_method ) {
+	public function get_title( $payment_method, $is_blocks = false ) {
+		if ( $is_blocks ) {
+			return $this->get_i18n( 'title_blocks_' . $payment_method );
+		}
+
 		return $this->get_i18n( 'title_' . $payment_method );
 	}
 
@@ -394,10 +400,15 @@ class AlmaSettings {
 	 * Gets title for a payment method.
 	 *
 	 * @param string $payment_method The payment method.
+	 * @param string $is_blocks Are we in blocks.
 	 *
 	 * @return string
 	 */
-	public function get_description( $payment_method ) {
+	public function get_description( $payment_method, $is_blocks = false ) {
+		if ( $is_blocks ) {
+			return $this->get_i18n( 'description_blocks_' . $payment_method );
+		}
+
 		return $this->get_i18n( 'description_' . $payment_method );
 	}
 
@@ -936,10 +947,11 @@ class AlmaSettings {
 	/**
 	 * Get eligible plans keys for current cart.
 	 *
-	 * @param array $cart_eligibilities The eligibilities.
+	 * @param array       $cart_eligibilities The eligibilities.
+	 * @param string|null $gateway_id The gateway id.
 	 * @return array
 	 */
-	public function get_eligible_plans_keys_for_cart( $cart_eligibilities = array() ) {
+	public function get_eligible_plans_keys_for_cart( $cart_eligibilities = array(), $gateway_id = null ) {
 		$alma_plan_builder = new PlanBuilderHelper();
 		if ( empty( $cart_eligibilities ) ) {
 			$cart_eligibilities = $this->get_cart_eligibilities();
@@ -956,7 +968,7 @@ class AlmaSettings {
 			}
 		);
 
-		return $alma_plan_builder->order_plans( $eligibilities );
+		return $alma_plan_builder->order_plans( $eligibilities, $gateway_id );
 	}
 
 	/**