+++ src/includes/Helpers/PluginHelper.php
@@ -18,6 +18,10 @@ use Alma\Woocommerce\Handlers\ProductHandler;
 use Alma\Woocommerce\Services\PaymentUponTriggerService;
 use Alma\Woocommerce\Services\RefundService;
 use Alma\Woocommerce\Services\ShareOfCheckoutService;
+use Alma\Woocommerce\Blocks\Standard\PayLaterBlock;
+use Alma\Woocommerce\Blocks\Standard\PayMoreThanFourBlock;
+use Alma\Woocommerce\Blocks\Standard\StandardBlock;
+use Alma\Woocommerce\Blocks\Standard\PayNowBlock;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit;
@@ -28,6 +32,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  */
 class PluginHelper {
 
+
 	/**
 	 * The order helper
 	 *
@@ -35,12 +40,20 @@ class PluginHelper {
 	 */
 	protected $order_helper;
 
+	/**
+	 * The block helper
+	 *
+	 * @var BlockHelper
+	 */
+	protected $block_helper;
+
 
 	/**
 	 * Constructor.
 	 */
 	public function __construct() {
 		$this->order_helper = new OrderHelper();
+		$this->block_helper = new BlockHelper();
 	}
 
 	/**
@@ -149,6 +162,38 @@ class PluginHelper {
 		// Launch the "share of checkout".
 		$share_of_checkout = new ShareOfCheckoutService();
 		add_action( 'init', array( $share_of_checkout, 'send_soc_data' ) );
+		if ( $this->block_helper->has_woocommerce_blocks() ) {
+			add_action(
+				'woocommerce_blocks_loaded',
+				array(
+					$this,
+					'alma_register_order_approval_payment_method_type',
+				)
+			);
+		}
+	}
+
+
+	/**
+	 * Register the blocks.
+	 *
+	 * @return void
+	 */
+	public function alma_register_order_approval_payment_method_type() {
+		// Hook the registration function to the 'woocommerce_blocks_payment_method_type_registration' action.
+		add_action(
+			'woocommerce_blocks_payment_method_type_registration',
+			function ( \Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry $payment_method_registry ) {
+				// Register an instance of Alma_Gateway_Blocks.
+				$payment_method_registry->register( new StandardBlock() );
+				$payment_method_registry->register( new PayNowBlock() );
+				$payment_method_registry->register( new PayLaterBlock() );
+				$payment_method_registry->register( new PayMoreThanFourBlock() );
+				$payment_method_registry->register( new \Alma\Woocommerce\Blocks\Inpage\PayNowBlock() );
+				$payment_method_registry->register( new \Alma\Woocommerce\Blocks\Inpage\InPageBlock() );
+				$payment_method_registry->register( new \Alma\Woocommerce\Blocks\Inpage\PayLaterBlock() );
+			}
+		);
 	}
 
 
@@ -158,7 +203,9 @@ class PluginHelper {
 	 * @return void
 	 */
 	public function wp_enqueue_scripts() {
-		if ( is_checkout() ) {
+		if (
+			is_checkout()
+		) {
 			$settings = new AlmaSettings();
 
 			$this->enqueue_checkout_scripts();
@@ -194,14 +241,25 @@ class PluginHelper {
 	protected function enqueue_in_page_scripts() {
 		wp_enqueue_script( 'alma-checkout-in-page-cdn', ConstantsHelper::ALMA_PATH_CHECKOUT_CDN_IN_PAGE_JS, array(), ALMA_VERSION, true );
 
-		$alma_checkout_in_page_js = AssetsHelper::get_asset_url( ConstantsHelper::ALMA_PATH_CHECKOUT_IN_PAGE_JS );
-		wp_enqueue_script( 'alma-checkout-in-page', $alma_checkout_in_page_js, array( 'jquery', 'jquery-ui-core' ), ALMA_VERSION, true );
+		if ( ! $this->block_helper->has_woocommerce_blocks() ) {
+			$alma_checkout_in_page_js = AssetsHelper::get_asset_url( ConstantsHelper::ALMA_PATH_CHECKOUT_IN_PAGE_JS );
+			wp_enqueue_script(
+				'alma-checkout-in-page',
+				$alma_checkout_in_page_js,
+				array(
+					'jquery',
+					'jquery-ui-core',
+				),
+				ALMA_VERSION,
+				true
+			);
 
-		wp_localize_script(
-			'alma-checkout-in-page',
-			'ajax_object',
-			array( 'ajax_url' => admin_url( 'admin-ajax.php' ) )
-		);
+			wp_localize_script(
+				'alma-checkout-in-page',
+				'ajax_object',
+				array( 'ajax_url' => admin_url( 'admin-ajax.php' ) )
+			);
+		}
 	}
 
 	/**